I don't remember what I changed

+++ 3ds/games/draw/upload.php
@@ -1,4 +1,5 @@
 <?php
+error_reporting(E_ALL);
 session_start();
 if(isset($_POST["i"]) && isset($_POST["t"]) && isset($_SESSION["ts_user"])){
 	$data = $_POST["i"];
@@ -8,14 +9,16 @@ if(isset($_POST["i"]) && isset($_POST["t"]) && isset($_SESSION["ts_user"])){
 		$id = rand(1000,3999);
 		$arry = array("from"=>$_SESSION["ts_user"],"time"=>time(),"src"=>"default","title"=>$_POST["t"]);
 		$aencode = json_encode($arry);
-		//if(!(file_exists("data/".$id.".json"))){
+		if(!(file_exists("data/".$id.".json"))){
 			file_put_contents("data/".$id.".json",$aencode);
 			list($type, $data) = explode(';', $data);
 			list(, $data)      = explode(',', $data);
-			$data = base64_decode($data);
+			$png = base64_decode($data);
 
-			file_put_contents('out/'.$id.'.png', $data);
-		//}
+			file_put_contents('out/'.$id.'.png', $png);
+		} else {
+			echo "file_exists";
+		}
 	} else {
 		echo "title2long";
 	}
